+++ tsup.config.ts
@@ -9,7 +9,8 @@ export default defineConfig({
   sourcemap: true,
   clean: true,
   minify: false,
-  treeshake: true,
+  treeshake: false,
+  cjsInterop: true,
   target: 'es2022',
   outDir: 'dist',
 });